feat!: remove legacy video editor methods from video block
